Remove leading slash for gRPC Web
so that it can respect the path prefix in the URI.

The current behavior is that the path is completely ignored because
there is a leading slash when the dart code is generated.

For example, '/protos.chat_service.ChatService/CreateStream'.

However, when it gets resolved using `Uri.resolve`,

```dart
void main() {
  final path = '/abc/def';
  final uri = Uri.parse('http://localhost:6060/grpc/');
  print(uri.resolve(path)); // http://localhost:6060/abc/def
  // /grpc/ is gone
}
```

By removing the leading slash when xhr_transport, it can work with path.
Only gRPC web supports path.

- fixes #258