Add simple end-to-end gRPC-web test
We did not have any before which allowed for regressions like #306 to
slip through. Unfortunately we can't test gRPC-web implementation
in pure Dart because we don't have a server side implementation of
the protocol. Instead we add a dependency on the third party
gRPC-web proxy (grpcwebproxy by Improbable - the choice made
purely based on the simplicity of installation) which forwards all
request gRPC server (written in Dart).