Fix for Windows Travis builds (#359)
This works around the issue with googlechrome package.

This package downloads an installer from official Google build source
which provides only the latest version. Whenever a new build is 
published Chocolatey package needs to be updated to resolve 
hash mismatch between new installer and what is hardcoded into the
package itself. Usually this happens automatically but
currently the package is stuck in moderation breaking
normal installation flow. 

We work around the issue by hardcoding the hash for the
current installer version.